accountoverzicht met middleware update

+++ routes/web.php
@@ -19,7 +19,7 @@ route::get('/about', function () {
     return view('about');
 });
 
-// Route::get('/home', [HomeController::class, 'index'])->middleware(['auth', 'verified'])->name('dashboard');
+ Route::get('/home', [HomeController::class, 'index'])->middleware(['auth', 'verified'])->name('dashboard');
 
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
@@ -32,10 +32,10 @@ Route::get('/admin', function () {
 })->middleware(['auth', 'verified', CheckAdmin::class])->name('admin');
 
 
-Route::get('/user/index', [UserController::class, 'index'])->name('user.index');
-Route::get('/user/edit/{id}', [UserController::class, 'edit'])->name('user.edit');
-Route::patch('/user/update/{id}', [UserController::class, 'update'])->name('user.update');
-Route::delete('/user/destroy/{id}', [UserController::class, 'destroy'])->name('user.destroy');
+Route::get('/user/index', [UserController::class, 'index'])->middleware(['auth', 'verified', CheckAdmin::class])->name('user.index');
+Route::get('/user/edit/{id}', [UserController::class, 'edit'])->middleware(['auth', 'verified', CheckAdmin::class])->name('user.edit');
+Route::patch('/user/update/{id}', [UserController::class, 'update'])->middleware(['auth', 'verified', CheckAdmin::class])->name('user.update');
+Route::delete('/user/destroy/{id}', [UserController::class, 'destroy'])->middleware(['auth', 'verified', CheckAdmin::class])->name('user.destroy');
 
 require __DIR__ . '/auth.php';
